Move question locator to treasure hunt package
